feat: Support new task format from YAML
The application can now parse a new YAML format for tasks where the tasks are nested under a "tasks" key.

This change is backward-compatible and the application will still be able to read the old format (a simple list of tasks).

Also, refactored the code to rename SampleTaskList to TaskList.